feat: support external pool `ACXwstETH` for staking
