+++ src/components/IconPair/IconPair.tsx
@@ -2,39 +2,38 @@ import styled from "@emotion/styled";
 import React from "react";
 
 type Props = {
-  MainIcon: React.ReactElement;
-  SmallIcon?: React.ReactElement;
+  LeftIcon: React.ReactElement;
+  RightIcon: React.ReactElement;
 };
 
 export function IconPair(props: Props) {
   return (
     <Container>
-      <MainIconContainer>{props.MainIcon}</MainIconContainer>
-      {props.SmallIcon && (
-        <SmallIconContainer>{props.SmallIcon}</SmallIconContainer>
-      )}
+      <LeftIconContainer>{props.LeftIcon}</LeftIconContainer>
+      <RightIconContainer>{props.RightIcon}</RightIconContainer>
     </Container>
   );
 }
 
 const Container = styled.div`
   position: relative;
+
+  img,
+  svg {
+    border: 2px solid #2d2e33ff;
+    border-radius: 50%;
+  }
 `;
 
-const MainIconContainer = styled.div`
+const LeftIconContainer = styled.div`
   height: 32px;
   width: 32px;
 `;
 
-const SmallIconContainer = styled.div`
+const RightIconContainer = styled.div`
   position: absolute;
-  right: -8px;
-  bottom: -1px;
-  height: 18px;
-  width: 18px;
-
-  svg {
-    border: 2px solid #3e4047;
-    border-radius: 50%;
-  }
+  right: -18px;
+  bottom: 0px;
+  height: 32px;
+  width: 32px;
 `;
