+++ src/utils/constants.ts
@@ -256,13 +256,35 @@ export type TokenInfo = {
   symbol: string;
   decimals: number;
   logoURI: string;
+  logoURIs?: [string, string];
   // tokens require a mainnet address to do price lookups on coingecko, not used for anything else.
   mainnetAddress?: string;
   // optional display symbol for tokens that have a different symbol on the frontend
   displaySymbol?: string;
 };
-// enforce weth to be first so we can use it as a guarantee in other parts of the app
 export type TokenInfoList = TokenInfo[];
+
+export type ExternalLPTokenList = Array<
+  TokenInfo & {
+    provider: string;
+  }
+>;
+
+export const externalLPsForStaking: ExternalLPTokenList = [
+  {
+    name: "Balancer 50wstETH-50ACX",
+    symbol: "ACXwstETH",
+    displaySymbol: "ACXwstETH",
+    decimals: 18,
+    mainnetAddress: "0x36Be1E97eA98AB43b4dEBf92742517266F5731a3",
+    logoURI: balLogo,
+    provider: "balancer",
+    logoURIs: [
+      acxLogo,
+      "https://assets.coingecko.com/coins/images/18834/small/wstETH.png?1633565443",
+    ],
+  },
+];
 
 // Order of this map determines the order of the tokens in the token selector
 export const orderedTokenSymbolLogoMap = {
@@ -283,38 +305,39 @@ export const orderedTokenSymbolLogoMap = {
   BOBA: bobaLogo,
 };
 
-export const tokenList: TokenInfoList = Object.entries(
-  orderedTokenSymbolLogoMap
-).flatMap(([symbol, logoURI]) => {
-  // NOTE: Handle edge case for Arbitrum and USDC combination.
-  // We currently do not support native USDC on Arbitrum, only the bridged USDC.e token.
-  // Until we do, we need to add a special case for USDC.e to the token list.
-  if (symbol === "USDC.e") {
-    const usdcTokenInfo = constants.TOKEN_SYMBOLS_MAP.USDC;
+export const tokenList = [
+  ...Object.entries(orderedTokenSymbolLogoMap).flatMap(([symbol, logoURI]) => {
+    // NOTE: Handle edge case for Arbitrum and USDC combination.
+    // We currently do not support native USDC on Arbitrum, only the bridged USDC.e token.
+    // Until we do, we need to add a special case for USDC.e to the token list.
+    if (symbol === "USDC.e") {
+      const usdcTokenInfo = constants.TOKEN_SYMBOLS_MAP.USDC;
+      return {
+        ...usdcTokenInfo,
+        logoURI,
+        symbol: "USDC.e",
+        displaySymbol: "USDC.e",
+        mainnetAddress: usdcTokenInfo.addresses[constants.CHAIN_IDs.MAINNET],
+      };
+    }
+
+    const tokenInfo =
+      constants.TOKEN_SYMBOLS_MAP[
+        symbol as keyof typeof constants.TOKEN_SYMBOLS_MAP
+      ];
+
+    if (!tokenInfo) {
+      return [];
+    }
+
     return {
-      ...usdcTokenInfo,
+      ...tokenInfo,
       logoURI,
-      symbol: "USDC.e",
-      displaySymbol: "USDC.e",
-      mainnetAddress: usdcTokenInfo.addresses[constants.CHAIN_IDs.MAINNET],
+      mainnetAddress: tokenInfo.addresses[constants.CHAIN_IDs.MAINNET],
     };
-  }
-
-  const tokenInfo =
-    constants.TOKEN_SYMBOLS_MAP[
-      symbol as keyof Omit<typeof orderedTokenSymbolLogoMap, "USDC.e">
-    ];
-
-  if (!tokenInfo) {
-    return [];
-  }
-
-  return {
-    ...tokenInfo,
-    logoURI,
-    mainnetAddress: tokenInfo.addresses[constants.CHAIN_IDs.MAINNET],
-  };
-});
+  }),
+  ...externalLPsForStaking,
+];
 
 // process.env variables
 export const gasEstimationMultiplier = Number(
@@ -416,12 +439,12 @@ export function getChainInfo(chainId: number): ChainInfo {
 
 export const tokenTable = Object.fromEntries(
   tokenList.map((token) => {
-    return [token.symbol, token];
+    return [token.symbol.toUpperCase(), token];
   })
 );
 
 export const getToken = (symbol: string): TokenInfo => {
-  const token = tokenTable[symbol];
+  const token = tokenTable[symbol.toUpperCase()];
   assert(token, "No token found for symbol: " + symbol);
   return token;
 };
