+++ src/views/Staking/components/StakingExitAction/StakingExitAction.tsx
@@ -1,20 +1,38 @@
 import BreadcrumbV2 from "components/BreadcrumbV2";
-import { Text, Logo, TitleLogo } from "./StakingExitAction.styles";
+import { IconPair } from "components/IconPair";
+
+import {
+  Text,
+  Logo,
+  TitleLogo,
+  IconPairContainer,
+} from "./StakingExitAction.styles";
 
 type StakingExitActionAttributes = {
   poolName: string;
   poolLogoURI: string;
+  poolLogoURIs?: [string, string];
 };
 
 export const StakingExitAction = ({
   poolName,
   poolLogoURI,
+  poolLogoURIs,
 }: StakingExitActionAttributes) => (
   <BreadcrumbV2
     onlyRootAncestor
     customCurrentRoute={
       <TitleLogo>
-        <Logo src={poolLogoURI} />
+        {poolLogoURIs ? (
+          <IconPairContainer>
+            <IconPair
+              LeftIcon={<Logo src={poolLogoURIs[0]} />}
+              RightIcon={<Logo src={poolLogoURIs[1]} />}
+            />
+          </IconPairContainer>
+        ) : (
+          <Logo src={poolLogoURI} />
+        )}
         <Text>Pool ({poolName})</Text>
       </TitleLogo>
     }
