+++ src/views/Staking/components/StakingInputBlock/StakingInputBlock.tsx
@@ -9,11 +9,13 @@ import {
   InputWrapper,
   MaxButton,
   TokenIcon,
+  IconPairContainer,
 } from "./StakingInputBlock.styles";
 import { capitalizeFirstLetter } from "utils/format";
 import BouncingDotsLoader from "components/BouncingDotsLoader";
 import { trackMaxButtonClicked } from "utils";
 import { useAmplitude } from "hooks";
+import { IconPair } from "components/IconPair";
 
 interface Props {
   value: string;
@@ -22,6 +24,7 @@ interface Props {
   invalid: boolean;
   buttonText: string;
   logoURI: string;
+  logoURIs?: [string, string];
   maxValue: string;
   onClickHandler: () => void | Promise<void>;
   displayLoader?: boolean;
@@ -37,6 +40,7 @@ const StakingInputBlock: React.FC<Props> = ({
   invalid,
   buttonText,
   logoURI,
+  logoURIs,
   maxValue,
   displayLoader,
   onClickHandler,
@@ -49,7 +53,16 @@ const StakingInputBlock: React.FC<Props> = ({
     <Wrapper>
       <InputRow>
         <InputWrapper valid={valid} invalid={invalid}>
-          <TokenIcon src={logoURI} />
+          {logoURIs ? (
+            <IconPairContainer>
+              <IconPair
+                LeftIcon={<TokenIcon src={logoURIs[0]} />}
+                RightIcon={<TokenIcon src={logoURIs[1]} />}
+              />
+            </IconPairContainer>
+          ) : (
+            <TokenIcon src={logoURI} />
+          )}
           <Input
             valid={valid}
             invalid={invalid}
